Revert "Feat/atatus instrumentation"
