+++ Engine/.gitignore
@@ -12,3 +12,4 @@ wheels/
 #Others
 close_msedge.py
 chromedriver.exe
+prototype.py
