+++ Engine/ClearRam.bat
@@ -1,0 +1,1 @@
+RAMMap.exe -Ew
