--- Engine/config/__init__.py
@@ -1,3 +1,0 @@
-from .config import CONFIG
-
-__all__: list[str] = ["CONFIG"]
