--- Engine/config/config.py
@@ -1,6 +1,0 @@
-
-CONFIG: dict = {
-    "ENV": "development", # Valid values: ["development", "production"]
-}
-
-__all__: list[str] = ["CONFIG"]
