+++ Site/package.json
@@ -14,7 +14,6 @@
         "@emotion/styled": "^11.14.1",
         "@mui/icons-material": "^7.2.0",
         "@mui/material": "^7.2.0",
-        "atatus-spa": "^4.6.5",
         "firebase": "^11.10.0",
         "lucide-react": "^0.525.0",
         "react": "^19.1.0",
