+++ Site/src/context/DownloadCountContext.jsx
@@ -2,8 +2,6 @@ import { createContext, useContext, useEffect, useState } from 'react';
 import { httpsCallable } from 'firebase/functions';
 import { functions } from '../config/firebase';
 
-import * as atatus from 'atatus-spa';
-
 
 const DownloadCountContext = createContext();
 
@@ -18,11 +16,11 @@ const DownloadCountProvider = ({ children }) => {
                 if (result?.data?.success) {
                     setDownloadCount(result.data.count);
                 } else {
-                    atatus.notify(new Error('Failed to fetch download count'), {receivedData: result.data}, ['download_count_fetch_error']);
+                    console.error('Failed to fetch download count:', result.data);
                 }
             })
             .catch((error) => {
-                atatus.notify(error, {}, ['download_count_fetch_error']);
+                console.error('Error fetching download count:', error);
             });
     };
 
@@ -32,11 +30,11 @@ const DownloadCountProvider = ({ children }) => {
                 if (result?.data?.success) {
                     setDownloadCount(result.data.count);
                 } else {
-                    atatus.notify(new Error('Failed to update download count'), {receivedData: result.data}, ['download_count_update_error']);
+                    console.error('Failed to update download count:', result.data);
                 }
             })
             .catch((error) => {
-                atatus.notify(error, {}, ['download_count_update_error']);
+                console.error('Error updating download count:', error);
             });
     };
