+++ Site/src/main.jsx
@@ -1,6 +1,3 @@
-import * as atatus from 'atatus-spa';
-atatus.config(import.meta.env.VITE_ATATUS_RUM_API_KEY).install();
-
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import App from './App.jsx'
