+++ UI/package.json
@@ -15,7 +15,6 @@
         "@mui/icons-material": "^7.1.1",
         "@mui/material": "^7.1.1",
         "@toolpad/core": "^0.16.0",
-        "atatus-spa": "^4.6.5",
         "firebase": "^11.10.0",
         "lucide-react": "^0.513.0",
         "react": "^18.3.1",
