+++ UI/src/Components/Cards/NewStorm/NewStorm.jsx
@@ -1,21 +1,19 @@
-import { useEffect, useState } from "react";
+import { useContext, useEffect, useRef, useState } from "react";
 import { useColorScheme } from '@mui/material/styles';
 import Card from '@mui/material/Card';
 import { CardHeader, CardContent, Divider } from "@mui/material";
 import { Settings2 } from 'lucide-react';
-import { logEvent } from "firebase/analytics";
 
 import "./NewStorm.css";
 import LeftPanel from "./Panels/Left/LeftPanel";
 import RightPanel from "./Panels/Right/RightPanel";
+import { CustomMUIPropsContext } from "../../../lib/ContextAPI";
 import ManageProfilesModal from "../../Modals/ManageProfiles/ManageProfiles";
 import fetchStatus from "../../../lib/FetchStatus";
 import { useStormData } from "../../../context/StormDataContext";
-import { analytics } from "../../../config/firebase";
-import { useCustomMUIProps } from "../../../context/CustomMUIPropsContext";
 
 const NewStorm = () => {
-    const { cardProps } = useCustomMUIProps();
+    const { cardProps } = useContext(CustomMUIPropsContext);
     const { colorScheme } = useColorScheme();
     const [manageProfilesOpen, setManageProfilesOpen] = useState(false);
     const formControls = useStormData();
@@ -31,12 +29,6 @@ const NewStorm = () => {
         }, 2000);
         return () => clearInterval(interval);
     }, []);
-
-    useEffect(() => {
-        if(manageProfilesOpen) {
-            logEvent(analytics, "manage_profiles_open");
-        }
-    }, [manageProfilesOpen]);
 
     return (
         <Card
