+++ UI/src/Components/Dialogs/AddChannels.jsx
@@ -1,21 +1,16 @@
-import { useEffect, useState } from "react";
+import { useContext, useEffect, useState } from "react";
 import { Avatar, Button, Checkbox, Dialog, DialogActions, DialogContent, DialogContentText, DialogTitle, LinearProgress, List, ListItem, useColorScheme } from "@mui/material";
-import { logEvent } from "firebase/analytics";
-
-import * as atatus from "atatus-spa";
 
 import "./Dialogs.css";
+import { CustomMUIPropsContext, SystemInfoContext } from "../../lib/ContextAPI";
 import { useLocalStorageState } from "@toolpad/core/useLocalStorageState";
 import { useNotifications } from "@toolpad/core/useNotifications";
 import ErrorText from "../Elements/ErrorText";
-import { useCustomMUIProps } from "../../context/CustomMUIPropsContext";
-import { useSystemInfo } from "../../context/SystemInfoContext";
-import { analytics } from "../../config/firebase";
 
 const AddChannels = ({ payload, open, onClose }) => {
     const { mode, defaultSelectedChannels } = payload;
-    const { btnProps } = useCustomMUIProps();
-    const systemInfoControls = useSystemInfo();
+    const { btnProps } = useContext(CustomMUIPropsContext);
+    const systemInfoControls = useContext(SystemInfoContext);
 
     const { colorScheme } = useColorScheme();
     const [hostAddress] = useLocalStorageState('hostAddress');
@@ -48,8 +43,7 @@ const AddChannels = ({ payload, open, onClose }) => {
             })
             .catch((error) => {
                 setErrorText("Failed to fetch channel data.");
-                atatus.notify(error, {}, ['channel_data_fetch_error']);
-                logEvent(analytics, "channel_data_fetch_error");
+                console.error("Error fetching channel data:", error);
             })
             .finally(() => {
                 setChannelsDataLoading(false);
