+++ UI/src/Components/Elements/CloseButton.jsx
@@ -1,4 +1,3 @@
-import { Fragment } from 'react';
 import { IconButton, Tooltip } from '@mui/material';
 import CloseIcon from '@mui/icons-material/Close';
 import { useColorScheme } from '@mui/material/styles';
@@ -8,15 +7,13 @@ const CloseButton = ({ onClick }) => {
     const { colorScheme } = useColorScheme();
 
     return (
-        <Tooltip title="Close" placement="bottom">
-            <Fragment>
-                <IconButton
-                    sx={{ position: 'absolute', right: '5px', top: '5px', '&:hover': { backgroundColor: '#ffffff10' } }}
-                    onClick={onClick}
-                >
-                    <CloseIcon sx={{ color: colorScheme === 'light' ? 'var(--dark-text)' : 'var(--light-text)' }} />
-                </IconButton>
-            </Fragment>
+        <Tooltip title="Close" placement="bottom"> 
+            <IconButton
+                sx={{ position: 'absolute', right: '5px', top: '5px', '&:hover': { backgroundColor: '#ffffff10' } }}
+                onClick={onClick}
+            >
+                <CloseIcon sx={{ color: colorScheme === 'light' ? 'var(--dark-text)' : 'var(--light-text)' }} />
+            </IconButton>
         </Tooltip>
     );
 };
