+++ UI/src/Components/Sections/Header/Header.jsx
@@ -5,10 +5,9 @@ import LightModeIcon from '@mui/icons-material/LightMode';
 import DarkModeIcon from '@mui/icons-material/DarkMode';
 import { Tornado } from 'lucide-react';
 import { useLocalStorageState } from '@toolpad/core/useLocalStorageState';
-import { logEvent } from 'firebase/analytics';
 
 import TornadoDark from "../../../assets/tornado.png"
-import { analytics } from '../../../config/firebase';
+
 
 import './Header.css';
 
@@ -21,7 +20,6 @@ const Header = () => {
         const newColorScheme = colorScheme === 'light' ? 'dark' : 'light';
         setColorScheme(newColorScheme);
         setDefaultColorScheme(newColorScheme);
-        logEvent(analytics, "theme_switch", { theme: newColorScheme });
     }
 
     return (
