+++ UI/src/lib/AppProviders.jsx
@@ -1,31 +1,38 @@
+import { useMemo, useState } from "react";
 import { DialogsProvider } from '@toolpad/core/useDialogs';
 import { NotificationsProvider, } from '@toolpad/core/useNotifications';
 
-
+import { CustomMUIPropsContext, SystemInfoContext } from './ContextAPI';
+import { getCustomMUIProps } from "./CustomMUIProps";
+import fetchRAM from "./FetchRAM"
+import { RAM_PER_PROFILE } from "./Constants";
 import { StormDataProvider } from "../context/StormDataContext";
-import { SystemInfoProvider } from "../context/SystemInfoContext";
-import { CustomMUIPropsProvider } from "../context/CustomMUIPropsContext";
 
 
-const AppProviders = ({ children }) => {
+const AppProviders = ({ children, theme }) => {
+
+    const MUIProps = useMemo(() => getCustomMUIProps(theme), [theme]);
+
+    const [availableRAM, setAvailableRAM] = useState(null);
+    const systemInfoControls = { availableRAM, setAvailableRAM, fetchRAM, RAM_PER_PROFILE };
 
     return (
-        <CustomMUIPropsProvider>
-            <DialogsProvider>
-                <NotificationsProvider slotProps={{
-                    snackbar: {
-                        anchorOrigin: { vertical: 'bottom', horizontal: 'left' },
-                        autoHideDuration: 2500,
-                    },
-                }}>
-                    <StormDataProvider>
-                        <SystemInfoProvider>
+        <CustomMUIPropsContext.Provider value={MUIProps}>
+            <SystemInfoContext.Provider value={systemInfoControls}>
+                <DialogsProvider>
+                    <NotificationsProvider slotProps={{
+                        snackbar: {
+                            anchorOrigin: { vertical: 'bottom', horizontal: 'left' },
+                            autoHideDuration: 2500,
+                        },
+                    }}>
+                        <StormDataProvider>
                             {children}
-                        </SystemInfoProvider>
-                    </StormDataProvider>
-                </NotificationsProvider>
-            </DialogsProvider>
-        </CustomMUIPropsProvider>
+                        </StormDataProvider>
+                    </NotificationsProvider>
+                </DialogsProvider>
+            </SystemInfoContext.Provider>
+        </CustomMUIPropsContext.Provider>
     )
 }
