+++ UI/src/lib/ContextAPI.js
@@ -1,0 +1,7 @@
+import { createContext } from 'react';
+
+export const CustomMUIPropsContext = createContext();
+
+export const StormDataContext = createContext();
+
+export const SystemInfoContext = createContext();
