+++ UI/src/lib/FetchStatus.js
@@ -1,5 +1,3 @@
-import * as atatus from "atatus-spa"
-
 const fetchStatus = async (formControls) => {
     fetch(`${formControls.hostAddress}/engine-status`, {
         method: 'GET'
@@ -9,7 +7,7 @@ const fetchStatus = async (formControls) => {
         formControls.setStormInProgress(data.storm_in_progress);
     })
     .catch((error) => {
-        atatus.notify(error, {}, ['status_fetch_error']);
+        console.error('Error:', error);
     });
 }
