+++ UI/src/lib/StormControlsClass.js
@@ -1,8 +1,4 @@
-import { logEvent } from "firebase/analytics";
 import ValidateStormData from "./ValidateStormData";
-import { analytics } from "../config/firebase";
-
-import * as atatus from "atatus-spa"
 
 class StormControlsClass {
     constructor(hostAddress) {
@@ -18,16 +14,6 @@ class StormControlsClass {
         this.setMoreChannelsLoading = null;
     }
 
-    log_analytics(stormData) {
-        if (stormData.subscribe) { logEvent(analytics, "subscribe") }
-        if (stormData.subscribe_and_wait) { logEvent(analytics, "subscribe_and_wait") }
-        if (stormData.subscribe_and_wait_time ) { logEvent(analytics, "subscribe_and_wait_time", { time: stormData.subscribe_and_wait_time }) }
-        if (stormData.slow_mode != 5) { logEvent(analytics, "slow_mode_change", { time: stormData.slow_mode }) }
-        if (stormData.background) { logEvent(analytics, "background_load") }
-        logEvent(analytics, "channel_count", { count: stormData.channels.length });
-        logEvent(analytics, "channel_selection_mode", { mode: formControls.channelSelection });
-    }
-
     startStorm(formControls, systemInfoControls) {
 
         const dataValid = ValidateStormData(formControls, systemInfoControls);
@@ -35,9 +21,6 @@ class StormControlsClass {
         if (!dataValid) {
             return;
         }
-
-        const stormData = formControls.getStormData();
-        this.log_analytics(stormData);
 
         formControls.setLoading(true);
         this.setControlsDisabled(true);
@@ -47,7 +30,7 @@ class StormControlsClass {
             headers: {
                 'Content-Type': 'application/json',
             },
-            body: JSON.stringify(stormData),
+            body: JSON.stringify(formControls.getStormData()),
         })
             .then(response => response.json())
             .then(data => {
@@ -57,21 +40,18 @@ class StormControlsClass {
                     });
                     this.setControlsDisabled(false);
                     formControls.setStormInProgress(true);
-                    logEvent(analytics, "storm_started");
                 } else {
                     formControls.setErrorText(data.message || 'Request failed');
                     this.notifications.show("Failed to start storm", {
                         severity: 'error',
                     });
-                    logEvent(analytics, "storm_start_failed");
                 }
             })
             .catch(error => {
+                console.error('Error starting storm:', error);
                 this.notifications.show(error.message || 'An error occurred while starting the storm', {
                     severity: 'error',
                 });
-                atatus.notify(error, {}, ['storm_start_error']);
-                logEvent(analytics, "storm_start_error");
             })
             .finally(() => {
                 formControls.setLoading(false);
@@ -97,16 +77,13 @@ class StormControlsClass {
                     this.notifications.show('Storm stopped successfully!', {
                         severity: 'success',
                     });
-                    logEvent(analytics, "storm_stopped");
                 }
-
             })
             .catch(error => {
+                console.error('Error stopping storm:', error);
                 this.notifications.show('An error occurred while stopping the storm', {
                     severity: 'error',
                 });
-                atatus.notify(error, {}, ['storm_stop_error']);
-                logEvent(analytics, "storm_stop_error");
             })
             .finally(() => {
                 this.setStopping(false);
@@ -132,15 +109,13 @@ class StormControlsClass {
                     this.notifications.show('Storm paused successfully!', {
                         severity: 'success',
                     });
-                    logEvent(analytics, "storm_paused");
                 }
             })
             .catch(error => {
+                console.error('Error pausing storm:', error);
                 this.notifications.show('An error occurred while pausing the storm', {
                     severity: 'error',
                 });
-                atatus.notify(error, {}, ['storm_pause_error']);
-                logEvent(analytics, "storm_pause_error");
             })
             .finally(() => {
                 this.setPausing(false);
@@ -166,15 +141,13 @@ class StormControlsClass {
                     this.notifications.show('Storm resumed successfully!', {
                         severity: 'success',
                     });
-                    logEvent(analytics, "storm_resumed");
                 }
             })
             .catch(error => {
+                console.error('Error resuming storm:', error.name);
                 this.notifications.show('An error occurred while resuming the storm', {
                     severity: 'error',
                 });
-                atatus.notify(error, {}, ['storm_resume_error']);
-                logEvent(analytics, "storm_resume_error");
             })
             .finally(() => {
                 this.setResuming(false);
@@ -199,20 +172,17 @@ class StormControlsClass {
                     this.notifications.show('Each channel will start storming without waiting for others!', {
                         severity: 'success',
                     });
-                    logEvent(analytics, "dont_wait");
                 } else {
                     this.notifications.show(data.message || 'Failed to set dont-wait', {
                         severity: 'error',
                     });
-                    logEvent(analytics, "dont_wait_failed");
                 }
             })
             .catch(error => {
+                console.error('Error setting dont-wait:', error);
                 this.notifications.show('An error occurred while setting dont-wait', {
                     severity: 'error',
                 });
-                atatus.notify(error, {}, ['storm_dont_wait_error']);
-                logEvent(analytics, "dont_wait_error");
             })
             .finally(() => {
                 this.setDontWaitLoading(false);
@@ -240,20 +210,17 @@ class StormControlsClass {
                     });
                     setMessages(messages);
                     setMessagesString(messages.join('\n'));
-                    logEvent(analytics, "change_messages");
                 } else {
                     this.notifications.show(data.message || 'Failed to change messages', {
                         severity: 'error',
                     });
-                    logEvent(analytics, "change_messages_failed");
                 }
             })
             .catch(error => {
+                console.error('Error changing messages:', error);
                 this.notifications.show('An error occurred while changing messages', {
                     severity: 'error',
                 });
-                atatus.notify(error, {}, ['storm_change_messages_error']);
-                logEvent(analytics, "change_messages_error");
             })
             .finally(() => {
                 this.setChangeMessagesLoading(false);
@@ -279,25 +246,23 @@ class StormControlsClass {
                         severity: 'success',
                     });
                     setSlowMode(slowModeValue);
-                    logEvent(analytics, "change_slow_mode");
                 } else {
                     this.notifications.show(data.message || 'Failed to change slow mode', {
                         severity: 'error',
                     });
-                    logEvent(analytics, "change_slow_mode_failed");
                 }
             })
             .catch(error => {
+                console.error('Error changing slow mode:', error);
                 this.notifications.show('An error occurred while changing slow mode', {
                     severity: 'error',
                 });
-                atatus.notify(error, {}, ['storm_change_slow_mode_error']);
-                logEvent(analytics, "change_slow_mode_error");
             })
             .finally(() => {
                 this.setChangeSlowModeLoading(false);
                 this.setControlsDisabled(false);
             });
+
     }
 
     startMoreChannels(channels) {
@@ -317,20 +282,17 @@ class StormControlsClass {
                     this.notifications.show('More channels started successfully!', {
                         severity: 'success',
                     });
-                    logEvent(analytics, "start_more_channels");
                 } else {
                     this.notifications.show(data.message || 'Failed to start more channels', {
                         severity: 'error',
                     });
-                    logEvent(analytics, "start_more_channels_failed");
                 }
             })
             .catch(error => {
+                console.error('Error starting more channels:', error);
                 this.notifications.show('An error occurred while starting more channels', {
                     severity: 'error',
                 });
-                atatus.notify(error, {}, ['storm_start_more_channels_error']);
-                logEvent(analytics, "start_more_channels_error");
             })
             .finally(() => {
                 this.setMoreChannelsLoading(false);
