+++ UI/vite.config.js
@@ -4,7 +4,4 @@ import react from '@vitejs/plugin-react'
 // https://vite.dev/config/
 export default defineConfig({
   plugins: [react()],
-  build: {
-    sourcemap: true
-  }
 })
